streamline stylesheets and element layout #192
Tweak the design of how blocks are laid out and styled by adding a
border to the outermost div and by introducing a new class of div whose
function is simply to have padding so content does not run directly up
against div borders. This necessitates removing borders and padding from
individual elements.

Remove real <hr> tags in favor of "fake" <hr> tags which probably
violates some css purism but they are so much better behaved. This and
the change above make the css DRYer and the layout more transparent to a
developer.

Comment out some css and reduce some nesting of elements to see if
anything bad happens (so far not).

Comment out second search field from "content-navbar-collapse". This
should be totally removed as its functionality should be covered by the
search field above.

Change style of inlined-error-box from "visibility: hidden" to "display:
none" which I think is what was intended. In a future commit this
element should be styled in an external style sheet instead of given an
inline style.