Tweak hiding/showing scripts #192
"event_changes.js" now finds the row that shows the associated major
action by its id instead of by assuming it is the last row in the
table. The earlier design caused final rows to disappear while
experimenting with moving the "archive" and "done" buttons to a position
below the first fieldset. Relatedly, "remove_delete_button" uses a more
streamlined approach to achieving the desired behavior.