Rethink padding and margins #192
Put padding on the inside of the content div, used for all views. Use
top and bottom margins on the .hr class, used to subdivide the content
div. These changes replace the "padded" class which is now removed. The
design is cleaner and requires no extra nesting to achieve attractive
spacing of elements.