Tweak styles for smaller screens #192
Now, on smaller screens: content div expands to full width, gap between
content div and upper nav disappears, upper nav gains some l/r padding
so elements aren't jammed against the sides of the screen.

Also, remove div with class "navbar-header" from around logo, as this
was causing the logo to jump on different screen sizes and the top of
the page is not a navbar in any case.

Also, add id attribute to anchor tags around logo so that the element
can be styled from swoptact.css instead of relying on inline styling.