Add enter-moves-to-next-cell and auto-scroll #200
When editing values in a dynamically generated table, pressing the
"enter" key now not only saves the value the user has entered (which it
always did), but opens input for the next column in the same row (only
if the next column is empty) or, if the user is entering data in the
rightmost data column of the bottom row, a new row will be added and the
focus put on the leftmost element in the new row. If none of the above
conditions apply, pressing "enter" just saves the data in the current
cell.

Known Issues: Currently, the above chain of events under certain
conditions when editing the name field. If the name is a new name, it is
now added twice instead of once. If the name is selected by
autocomplete, values for the remaining fields are also filled in (is
this correct behavior?).

Also, the TAB key should have the same functions as the 'enter' key but
the tab key event is not captured by the "keyup" handler (it is caught
by a 'keydown' handler).