Add calendar to end date in search: #185
A couple things happened here.  One, we were missing a close >, which
was easy to fix (once I found it).  Then, it looks like we didn't
actually need to include the DateTimeShortcuts.js file on the page,
because it gets called from the AdminDateWidget in Django.  The actual
problem was that jQuery wasn't loading on the page as "django.jQuery,"
because we're not using Django's jQuery, but rather loading a library
from a file.  So, I added the "alias_jquery" file and included it in our
base template to avoid that issue in the future.  Now, both start and
end dates have a calendar datepicker for the advanced search.