Exclude archived objects from missing data: #196
Add a conditional to check whether an object has been archived before
including it in the results from the missing data report.  I only added
one line.  Most of this diff is indentation change.  @cwebber, does this
look like the right change?  Or should this change go in the arguments
to gather_results?