Move form tags to not include dynamic table #200
The form tags only properly belong around the top half of the
change_form template since the ajax-populated table is updated
automatically (without an explicit 'submit' command).

This demarcation is especially important so that when TAB or ENTER is
pressed, a keydown event handler can behave differently depending on
whether the focus is on the traditional form or on the dynamically
generated table.