Redesign and separate editing functionality #200
The implementation of new functionality whereby pressing the ENTER or
TAB key while editing a cell in an ajax-populated table would mean
"finish editin the cell and move on to the next one, starting a new row
if necessary" was running into trouble when new rows were added.  The
design before now had created an empty table row into which the user
would enter a value into the leftmost column.  Once this was entered, a
primary key id was retrieved from the DB (either for an existing entity
or a new one), and the row into which the user was typing was then
removed from the DOM.  The row was then replaced with a nearly identical
row which contained the id as received from the DB.

The destruction and recreation of rows made it unreasonably difficult to
do things like placing the user's cursor in the next field over.  Given
this, and the fact that other hacks to deal in a special way with table
rows which were "new" (i.e., lacking an id) were already present in the
code, a streamlining of the code seemed the best way to move forward.

Pursuit of this ideal led to the following changes:

Separation of the functions of creating DOM elements (i.e., table rows
and cells), populating those elements with data, creating profile links,
and appending rows to the "inlined" table.  Some of these steps were
previously grouped together, making customization for different
use-cases difficult.

Removal of the functions "setupStatic" and "setupEdit", which hooked
into one of two different methods depending on whether the cell to be
edited was connected to an autocomplete function.  The different
use-cases differed only slightly after the separation of creation and
population functionalities mentioned above, and simple if-then branching
now suffices.

The function "maybeUpdateNew" is now a freestanding function and no
longer part of one of the "setupEdit" use-cases mentioned above.

All event handlers are now defined at the end of the file.

In addition, the previous technique of handling a "blur" event on table
cells to trigger the saving of those cells' contents to the database has
been replaced by instead handling "click" events anywhere on the
document, which is easier to manage.