Improved click handling and profile links #200
Previously, if the user selected an auto-complete option in the "name"
column while editing in the bottom row, the row was filled in, a new row
was added to the bottom, and the first column of that new row was opened
for editing. This was accomplished by first programmatically raising a
click event on the "add new" button and then programmatically raising a
click event on the first cell of the new row. However, these two click
events in quick succession bubbled up to the document level where each
triggered the addition of a profile link (the "i" in a circle) to the
name column, resulting in a name column with two profile links. Now the
changes needed to open a cell for editing have been abstracted into a
new function which can be called either by the click handler or
independently, so that the above behavior is maintained with only a
single click event (raised on the "add new" button) bubbling to the
document.

Relatedly, multiple profile links could still occur if the user manually
did multiple clicks (or keypresses of ENTER/TAB) while editing a
cell. To correct this, the profile link is now added to every table cell
as it is constructed, but is only made visible when a real link
exists. This has mulitple advantages over creating and placing a new dom
element upon request as had previously been done: 1. No more than one
profile link will appear in a cell, 2. All dynamic construction of dom
elements for ajax-populated tables is now consolidated into the
"getNewTableRow" function. 3. The fact that the profile links are by
using "visibility: hidden" as opposed to "display: none" means they
continue to take up space in their table cells even when invisible,
which in turn means that the column widths do not change when switching
between editing and non-editing modes.

Known issues: profile links are now created (although they are not
visible to the user) in columns where none will ever exist, (e.g., in
the phone number column). This should probably be corrected by adjusting
the model in some way to indicate whether a given member can have an
associated profile link.