semi-correct jumping scroller #200
For reasons I haven't been able to figure out, the appearance of a
validation error message in a cell of the ajax-populated table causes
the table to scroll upwards so that the error is no longer in view. This
revision does not really correct the problem, but it does scroll the
table back down again so that although there is a perceptible flicker in
the table as it jumps up and back down, the error message is in view at
the end.

I could not find anything in the Swop code that could cause this.  It is
possible that the problem is in the "sticky header" javascript plug-in.