No styling for cells in same row as an error #200
A validation error in a cell of the ajax-populated table causes a red
border to appear around that cell.  In addition, other cells in the same
row got a red/pink background both to draw further attention to the row
and also to serve as a visual reminder that those cells cannot be edited
until the error in the sibling cell is corrected.  (An alert-box appears
if the user attempts to edit these cells).

The red/pink background has now been removed, as this could have implied
that errors existed in those cells as well.  However, the alert box
warning still appears if the user attempts to edit them.