Fix submit bug #210
The submit button in "change_form.html" (labeled "done") was no longer
between the opening and closing form tags in after these tags were
repositioned in commit 520bb44. The tags were repositioned to exclude
the ajax-populated table in the lower section, which is not part of the
main form as its values are saved immediately after entry and do not
require a specific "submit" action.  Furthermore, the ajax-populated
table should also be placed outside the form tags because the table and
the main form each have their own handlers for TAB and ENTER keypress
events that lead to different behavior.

However, for aesthetic reasons, the submit button continues to be placed
at the bottom of the page, outside the form tags. Since it was no longer
located between the form tags, it became disconnected from the form, and
hence we need a new handler ("handle_submit_button.js") to reconnect it.