Properly fix jumping table #200
Commit 16fd6cd took a naive stab at this problem, but this is now the
correct fix. The problem is that adding an error message into a cell may
change the cell's dimensions (i.e., make it wider), which may change the
dimensions of other cells (i.e., make them narrower). When this happens,
the errored cell may be forced below the constant-height viewport that
is part of the sticky table headers implementation. Although inserting a
validation error produced an illusion that the table was scrolling up,
in fact no scrolling occurred at all; rather, the heights of the table
rows above the error were increasing while the viewport remained
stationary, leaving the user looking at a different part of the table.

Therefore, this revision makes sure the distance between the top of the
errored cell and the top of the viewport remains constant. Adding an
explicit position rule to class "scrollable-area" allows us to use it as
a reference point for measuring the position of a table cell.