Move messages inside div #202
Now all messages appear as do error messages, which already appeared
attractively at the top of the main div in a colored box above the
title. The color of the box is determined by the "tags" property on the
message. Furthermore, if there is only one message to be printed, it
will not be printed in list format (with a bullet preceding it), but
simply as a sentence (as error messages are).