Vertical filter for group permissions #202
Following commits e341a76 and 6d462b9, the Django "filter_vertical"
widget became the tool for assigning users to groups instead of the
previous "filter_horizontal". This commit does the same thing for
assigning permissions to groups. It also further massages the styles of
the filter widget to sit comfortably inside the main div and play well
with other visual elements.