Remove double instance of page title #202
Removing the "{% block content_title %}" from around the page title
markup in base.html eliminates the occurrence of having the title appear
twice on the user's admin screen (accessible by clicking "swoptact" in
the breadcrumb trail).  The commit also removes markup to display the
table from various other templates, which should be unnecessary since
all templates ultimately inherit from "base.html".

Note that there is also a "{% block title %}" in "base.html" which
appears in the head of the document between the html <title> tags. Thsi
block only appears in one other template, "base_site.html", in which it
appears in the body of the markup. Has a mistake been made somewhere?