Fix problem introduced in 2171350
In 2171350 a problem was introduced where regardless of filters
it always showed participants who weren't part of any event or
institution. This now checks if any filters have been applied
narrowing it down by either events or institutions and if so
it ignores those users who are not related to those.

Touch #185