Simplified version of #185
Remove some advanced search options and change labels.  This version
only allows searching for participants, but in the next commit I'll
return the option to search for actions.  I wanted to simplify it so
that I could be sure that all search combinations worked and so that the
UI would be a little simpler.  I might deploy this simple version to
beta for feedback and then add the action options.