Return action search option #185
Allow searching for actions (I had removed that option in 724ea81).
Show and hide search fields according to whether the user is searching
for participants or for actions.  Alter field labels from their original
state.

The next things I'll add are a clear filters button and possibly the
ability to search actions by "connected action."  I'm not sure how
difficult it would be to alter the search query to include that.