Change background and small fixes: #246 & #254
Add a new background image, concrete_seamless.png, from
www.subtlepatterns.com.  The copyright notice on that site says "All
patterns on this site can be used free of charge, but please read this
before using them. CC BY-SA 3.0 - Subtle Patterns © Atle Mo."

Show the "i" in a circle for the institution link (make the "large-info"
class available outside the inlined-model-table, so that the circle
shows up correctly on the participant change form).  Change the is_prep
label.  Remove the link to the missing data report.