Font consistency changes: #245
I'm going to need a lot of feedback on this from the users and @kfogel.
For now: I've removed the different header color and font, so that the
only colors we have are purple, black, and white and all fonts are
sans-serif (Arial).  I think the <h1>s are a little heavy now.  I
changed links to underline on hover instead of losing their underline on
hover, so that the space is less cluttered.  Maybe we want purple
headers instead of black?