Start #247: add leader stages and UI
This adds the four possible leader stages and an inlined interface for
editing them on the participant profile.  However (I think because we've
created a special form for saving Participants) these don't yet save.
The next step is to fix that, and then to create permissions so that
this section is only visible for staff and admin users.