Show filtering options as dropdown: #250
Instead of an displaying an endlessly growing list of tags or other
filters, show them as a dropdown.  To do this, I used a custom template
for the filters and displayed them as a "select" element instead of an
unordered list.  Since they were no longer in anchor tags, I needed to
write some javascript to reload the page on change of any select element
of this "dropdown_filter" class.

The next step (the only remaining step in this issue?) is to abstract
the "TagFilter" class so that it can handle other filters we might want
to use.  The most immediate are the filters on the user list: "active"
and "group."