Reorganize PUT/POST requests into functions: #268
Instead of having PUT and POST requests in a conditional in the
saveInlinedModel() function, move them each into their own function.  In
our new nonce-using structure, we need to send PUT requests in two
different branches of our conditional tree, so I need to be able to call
the same function in two different places.  Clean up the conditional
tree a little bit.