Create the nonce_to_id table and save: #268
On POST, save the nonce and new participant ID to the now-in-existence
nonce_to_id table.  (Note: this commit includes a migration.  Run python
manage.py migrate to create the new table.)  As I noted in a comment,
I'm unexpectedly getting new nonce records with each PUT, too.  I think
the put() and post() methods in APIMixin must be interacting in some way
that I don't yet understand.

As of this commit, new participants are created successfully and their
nonces are saved.  Next is to handle the _real_ bug, which occurs when a
PUT should be sent but the client doesn't have an ID for the participant
yet.