Check changed fields on client side: #273
Determine which fields were changed and send those (with their new
values) along with the old value of the stored object.  This results in
breakage on the server side, since I haven't made changes there yet to
accomodate this new information (that is, this does not work as of this
commit!).  Note also that the PUT URL constructed by the client still
expects an ID, which it may not get.

Remove @cwebber's comment from 4e1a9d2 because these changes are
designed to alleviate the danger he was warning about.

Next is to change the server's PUT endpoint so that it expects the new
inputs and can behave appropriately.