Pass changed fields and values: #273
Send an object with old and new values for any fields that were changed,
instead of the entire object.  Make one attempt at handling this
information on the server side, but this version is far overspecified
for institution (it works for institution, but not for phone number or
any other change).  Recording it here in this commit for posterity, but
I need to make a more universally applicable server-side flow.