No more duplicates and PUT working: #268, #273
POST and PUT requests can now be processed correctly without creating
duplicate participants from the sign-in sheet, even on a slow
connection.  Since a PUT is eventually processed by the post() function,
update the request.POST element with the changed fields and send that to
the post() function, which saves and returns to the page, so that the
user has the same experience as s/he did with the old system.

However, as of this commit this code still assumes that the object being
altered by PUT is a participant, which may not always be true.  Also,
new phone numbers entered via PUT are not processed correctly, so will
not save and will instead throw a 500 response to the page.

After fixing the problems in the previous paragraph, I need to add
proper error handling.  Right now I just have a lot of lines with
debugging statements as reminders to add errors.