PUT works for all fields: #268, #273
Institutions and phone numbers can both be added and edited now.  I
think I need a more global solution than the change in this commit, but
this works for now.  Besides errors, I need to handle the nonce-only PUT
case, which currently gives a 404 because it tries to PUT to
`/api/participants//` with no ID.