PUT to nonce-based endpoint: #268, #273
I can now PUT to a nonce-based endpoint and get a 200 response, as
expected.  However, institutions still do not save correctly (i.e., if I
create a new participant from the sign in sheet and add a brand-new
institution to that participant, the institution will disappear from the
sign-in sheet and will not be created in the database).  That's the next
thing to investigate.