Preserve in-progress client data: #268
When a POST or PUT request is finally fulfilled, the table row of that
request is re-filled with data.  Save and replace the initial contents
of the inputs in that row, so that users don't lose data (e.g., if a
user is entering a phone number we want it to remain there even when the
person's name is saved and the row is re-filled with the new person's
data from the server).

Also, remove an unnecessary error check from the server side.