Remove references to "swoptact"
Start renaming this project to StreetCRM.  Remove references to
swoptact.  There are still a lot of references to SWOPtact and SWOPTACT,
not to mention SWOP the organization.  More replacements coming soon.

Once I've changed the name everywhere and made sure that everything
still works, I'll make the logo and color scheme configurable.