Remove (most) references to "SWOP"
Remove references to SWOP, including in copyright headers.  A few
remain, because they will be a little more difficult to alter.  These
are in migration 30 and models.py, where one of the field labels
currently has "SWOP" as part of its text and will need to be
configurable, and the logo on the front page.

Remove an old notes file that is now obsolete.