Make org name in institution label configurable
Add an example/default name that isn't SWOP to config.example.ini.  Add
a migration that applies that default name to the label, and note in
INSTALL.md what needs to be done to change it to the actually desired
org name when the config file is updated.  (That is to say, if we wanted
to update the org name to OTS, we'd change it in the newly constructed
config file and then make and run the migration, as described in
INSTALL.md.)  Unfortunately, because of the way field labels work in
Django, I think we need to have a migration here, even though it's a bit
unwieldy.

Comments welcome, @kfogel et al.