Add possible migration for existing instances
The migration changes in 7b4f817 are all very well for new instances,
but we also need to account for existing instances.  This migration
(here as a .tmpl file to avoid it being automatically run by Django) is
an attempt to solve that problem.  For now it doesn't quite work.  That
is, it renames the tables correctly and the app works, but future
migrations will expect all previous migrations to have been run.  So,
the changes in 7b4f817 cause errors when I try to run new migrations,
even after renaming the tables with the migration in this commit.