Change color and clarify INSTALL.md
Change the default color in streetcrm.css to green (for OTS) instead of
purple.  This is not the right way to do that.  I'm only committing at
@kfogel's request, so that we know where we need to change the colors
once I learn how to make the color configurable.  Also note that
streetcrm/static/images/three_lines_purple.png is always purple.  We
need to make that image configurable like the logo.