Start work on searching for institutions: #280
Start two parts of this issue:
- include institutions in search results
- add more help text for search

Neither of these are working perfectly yet.  A comment in #244 suggests
that adding `search_fields` to InstitutionAdmin should make institutions
show up in results, but it doesn't.  This probably has to do with our
custom search infrastructure.

Current Watson docs suggest using the new AppConfig, which was bumped up
in priority with Django 1.9 (which came out after most work on StreetCRM
happened).  This commit starts setting that up, though it might not be
the right path forward since we already have quite a bit of custom code
written under the old system.