Add institutions to search results: #280
Restore all search_fields to admin.py.  Now I see how this works.  For
each model that has a search_fields entry in admin.py, when a new
instance of that model is created another entry is created in the
`watson_searchentry` table.  When we run a basic search, the query is
compared against entries in that searchentry table.

So, new institutions after this commit will show up in search results.
I'll need to write a script to add existing ones to the searchentries
(or find out the regular Watson way to handle existing objects when
registering a new model).