Return CSV using search query: #281
Call search again and return a CSV after clicking a link in the view.
This does not yet use the right search query (see where I set the search
query in views.py), but it does everything else that it should to
generate a csv of search results.  Next, pass the search query
correctly.  Then, clean up the exported file.

This code still includes the old URL and view for doing exports as a
POST.  That needs to be cleaned up.