Make export icon clickable to fix #292
Add export icon (which is a class that comes with bootstrap) as a label
for the "input" element so that it and the text are both clickable and
appear as one button.

Thanks to [answers to this StackOverflow
question](http://stackoverflow.com/questions/9376192/add-icon-to-submit-button-in-twitter-bootstrap-2)
for pointers on how to do this.