Add migrations to rename tables and fix permissions
These migrations are necessary for upgrading an existing instance to
master, but should not themselves exist on master (master tables already
have the correct names and permissions).  After this upgrade is
complete, I'll move the instance to master and delete this branch.

I'm keeping the swoptact-safe-deploy branch separate for now so that it
can actually be a "safe-deploy" branch until we're sure that the upgrade
went smoothly.