Reorder objects on the front page of StreetCRM for #291
The app name was case-sensitive, so "StreetCRM" in settings was not
interpreted as the same app as "Streetcrm" elsewhere (I'm not sure
where).  Since the ADMIN_ORDERING value is only used in
custom_ordering.py, I changed the case to work with that file.  The
objects now appear in the order specified in settings.py