Write test for login: #297
Using the correct username and password, this test now passes.  Update
sample data to include the sample username and password, so that this
test will pass immediately on setup of a new instance.

One thing that is still unclear to me about Selenium is when we use the
BasePageElement class to locate an element as opposed to a locator.

Next, I'll chain tests so that I can test running a basic search after
logging in.  The basic search test is currently commented out (in
progress) so that the login test passes.