Show "see more" before 6th action: #298
This change ensures that the "see more" link shows up only if there are
in fact more actions to show for a given participant.  Before the
commit, the "see more" text was appended to the 5th action for a given
participant.  This meant that for participants who had attended exactly
5 actions, the "see more" link showed up but had no effect.  Now, it
will be prepended to the 6th action, meaning it will only appear on the
page if the participant has attended at least 6 actions.