Stop the browser from autofilling the name field: #288
Turn the "autocomplete" attribute off so that the browser doesn't offer
suggestions for names.  Many of our fields have StreetCRM-specific
autocomplete, which will stay.  This just stops the browser from
offering suggestions for the names of new Participants, Actions, and
Institutions.

Note that both Actions and Institutions use the
TagSkippingAutoCompleteModelForm.