Grant admins permission to delete objects: #290
In addition to archiving, add ability to delete objects.  As of this
commit, archiving and deletion both work.  The buttons need to be
separated and better-styled.  More significantly, there is nothing in
the template to distinguish between who should see only the archive
button (users in the staff group) and those who should see the delete
button (admins and superusers)