Alphabetize participant results from advanced search: #305
This *breaks* action results from advanced search.  The next commit will
fix that and alphabetize those as well.

The method we use to do search returns different types (list vs dict)
for action and participant results.  My aim is to change that, so they
can both be sorted and displayed the same way.