Initial working CSV import, WIP
Updates the FileImportMixin to be general utilities for handling CSV
data, adds some sample files, establishes initial imports for Events,
Participants, and Event participants, and updates the initial UI for
imports.

NOTE: bulk_create was being used, but on Django 1.8 it does not return
primary keys. Additionally, the way the import button is included in
change list templates needs to be improved.