Update toast message, hide on click
Moved the toast message to the center of the page so that it works for
more screen sizes. Also made it more clear for the sign-up page how many
participants were created and how many already existed but were added to
the event. Had to make the toast message a bit bigger for this reason,
so increased the time it's visible, but also added a click handler so
that someone can dismiss it before the end of the timeout.