Add ID, leadership stage to export
Adding ID is straightforward, but the query for leadership stage is now
updated to use the through model LeadershipGrowth because it tracks the
time. This is assuming that the export will pull the latest leadership
stage, and not necessarily the one that was most recent for the last
date of the query. Converting the prefetched queryset to a list, because
sorting it directly will trigger another DB call for each row