Require specifying all import_fields, handle ID
The FileImportMixin now requires that all fields that should be
imported or queried on import need to be specified on field_map. This is
so every field that we would want to ignore (like leadership_stage) can
be ignored by default, and to make the import assumptions more explicit.
Also, if ID is present, it will be returned as the only field to query,
handling the fact that it's now a part of the default export.