Fix attendance counting, export columns/headers
Add any search filters on event date to the count in participants search
so that events outside the filtered range aren't included. Add None to
leadership export row if growth steps not found to make sure the column
is created. Check for datetime and date types in search header creation.