Change Group import location in signals.
When booting up in django 2, there was an error message about Apps not
being loaded yet.  It was an error to load up Group from auth.models
before the appliaction had loaded, and so the import statement needed to
be moved into the function call, rather than at the top level which was
being initialized too early.