Comment out use of RadioFieldRenderer
Django removed some undocumented classes, but doesn't really provide an
upgrade path.  The correct thing to do is to figure out how it was used,
and then upgrade to whatever the future should be, but that was not
obvious (to me).  Instead, in the attempt to get the system booted, this
commit just comments it out, and replaces by the closest sounding
django class.

Probably going to break the application.