Fix #340: Breadcrumb links disappear on hover
Add a new rule for breadcrumb css to make a:hover white as well.  Added
the new rule instead of removing the rule that affected this on hover
because that rule is part of the theme code in base.html, and special
casing when a theme shouldn't be applied is better than special casing
when it should.